refactor(sweep): extract per-session reconcile; make the recurrence re-arm atomic
Code motion plus one durability fix, behavior otherwise byte-equivalent:

- The per-session body of the host sweep (ack sync, due-message wake,
  stuck/idle enforcement, mailbox maintenance) moves verbatim from
  host-sweep.ts into reconcile-session.ts as reconcileSession(sessionId) —
  the ReconcileFn shape from src/reconcile.ts. Level-triggered: it loads
  current state, and a missing or closed session is a clean no-op.
  host-sweep.ts keeps the 60s loop, the active-session iteration, and the
  once-per-tick duties, and re-exports the moved public surface so every
  existing import (and test) resolves unchanged.

- The recurring-task re-arm was two separate durable writes: insert the next
  occurrence, then clear the recurrence on the completed original. A crash
  between them re-clones the series on the following tick — duplicate runs.
  New InboundMailbox.armNextTask(originalId, task) does the pair as one
  transaction (the applyProcessingAcks idiom); both re-arm paths in
  handleRecurrence now use it. A failure mid-pair provably leaves the
  original armed — retryable, never dead, never duplicated.

Zero edits to existing tests: the sweep suites pass against the re-exported
surface, and the recurrence suite exercises the real SQLite mailbox, which
now implements the atomic pair.
